+++ R/forest.netsplit.R
@@ -192,6 +192,8 @@ forest.netsplit <- function(x,
   ##
   meta:::chkclass(x, "netsplit")
   ##
+  x <- upgradenetmeta(x)
+  ##
   chkchar <- meta:::chkchar
   chklogical <- meta:::chklogical
   chknumeric <- meta:::chknumeric
@@ -371,10 +373,11 @@ forest.netsplit <- function(x,
   else
     dat.predict$TE <- NA
   ##
-  dat.predict$seTE <- dat.predict$z <- dat.predict$p <- dat.predict$prop <- NA
+  dat.predict$seTE <- dat.predict$statistic <- dat.predict$p <-
+    dat.predict$z <- dat.predict$prop <- NA
   ##
   dat.predict <- dat.predict[, c("comparison", "TE", "seTE",
-                                 "lower", "upper", "z", "p", "prop")]
+                                 "lower", "upper", "statistic", "p", "prop")]
   ##
   dat.direct$comps <- dat.indirect$comps <-
     dat.overall$comps <- dat.predict$comps <- x$comparison