+++ R/formatComp.R
@@ -1,7 +1,7 @@
 formatComp <- function(x,
                        backtransf, sm, level,
                        trts, trts.abbr,
-                       digits, digits.zval, digits.pval.Q,
+                       digits, digits.stat, digits.pval.Q,
                        scientific.pval, big.mark) {
   
   
@@ -38,18 +38,24 @@ formatComp <- function(x,
                                        digits, "NA", big.mark),
                                formatN(round(res$upper, digits),
                                        digits, "NA", big.mark))
-  res$z <- formatN(res$z, digits.zval, big.mark = big.mark)
+  res$statistic <- formatN(res$statistic, digits.stat, big.mark = big.mark)
   res$p <- meta:::formatPT(res$p,
                            digits = digits.pval.Q,
                            scientific = scientific.pval)
   ##
-  res$upper <- NULL
+  res$upper <- res$z <- NULL
   ##
   sel <- names(res) == "TE"
   names(res)[sel] <- sm.lab
   ##
   sel <- names(res) == "lower"
   names(res)[sel] <- ci.lab
+  ##
+  sel <- names(res) == "statistic"
+  names(res)[sel] <- "z"
+  ##
+  sel <- names(res) == "p"
+  names(res)[sel] <- "p-value"
   ##
   colnames <- names(res)
   ##